moved blocktable and address info container to explorer
