--- packages/taraxa-ui/src/components/AddressInfo/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AddressInfo';
