--- packages/taraxa-ui/src/components/BlockTable/index.ts
@@ -1,1 +1,0 @@
-export { default } from './BlockTable';
