+++ services/explorer/src/components/AddressInfo/index.ts
@@ -1,0 +1,1 @@
+export * from './AddressInfo';
