+++ services/explorer/src/components/BlockTable/index.ts
@@ -1,0 +1,1 @@
+export * from './BlockTable';
