+++ services/explorer/src/components/index.ts
@@ -2,3 +2,5 @@ export * from './Footer';
 export * from './Header';
 export * from './icons';
 export * from './PageTitle';
+export * from './AddressInfo';
+export * from './BlockTable';
